requirements.txt
cycler==0.11.0
fonttools==4.37.1
kiwisolver==1.4.4
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
